TP3 - finished

+++ TP3/integration_gaussienne.m
@@ -6,32 +6,32 @@ function [ In ] = integration_gaussienne( type, nb_points, r )
 %1 -> Tchebyschev
 
     In = 0;
-    % On transoforme la fonction symbolique en fonction de matlab
+    % On transforme la fonction symbolique en fonction de matlab
     r = matlabFunction(r);
     switch type
     case 0
         % Integration Legendre
-        % On recupere les polynomes de Legendre pour le nombres de points
+        % On recupere les polynomes orthogonaux de Legendre pour le nombre de points
         % donne
         poly_legendre = polynome_legendre(nb_points);
-        % On alloue un peu de memoire pour contenir les racine des
+        % On alloue un peu de memoire pour contenir les racines des
         % polynomes
         zeros_legendre = zeros(nb_points-1, nb_points-1);
         for i=2:nb_points
             % On calcule les zeros des polynomes de Legendre
-            % solve n'est plus capable de trouver les racine a partire d'un
-            % polynome de degre 9. On utilise donc roots qui prend en parametre
-            % les coefficien du polynome, qui sont donnee par sym2poly
-            rootsPoly = double(roots(sym2poly((poly_legendre(i)))));
+            % solve n'est plus capable de trouver les racine a partir d'un
+            % polynome de degre 9. On utilise donc roots qui prend en parametres
+            % les coefficients du polynome, qui sont donnes par sym2poly
+            rootsPoly = roots(sym2poly((poly_legendre(i))));
             zeros_legendre(:, i-1) = [rootsPoly; zeros(nb_points-i, 1)];
         end
         % On s'interesse a la transpose de la matrice
         zeros_legendre = zeros_legendre';
-        % On calcules les coefficients de Legendre
+        % On calcule les coefficients de Legendre
         coefficients_legendre = coeff_legendre(poly_legendre, zeros_legendre);
         
         for i=1:nb_points-1
-            % On recupere la ieme racine des polynome de Legendre
+            % On recupere la ieme racine des polynomes de Legendre
             xi = double(zeros_legendre(nb_points-1, i));
             % On applique la formule 
             In = In + coefficients_legendre(nb_points-1, i)*r(xi);
@@ -42,7 +42,7 @@ function [ In ] = integration_gaussienne( type, nb_points, r )
         zeros_tchebyschev = zeros(nb_points-1, nb_points-1);
         % On calcule les polynomes de Tchebyschev
         poly_tchebyschev = polynome_tchebyschev(nb_points);
-        % On calcule les racines des polynomes. On n'utilise pas solve pour les meme raison que pour
+        % On calcule les racines des polynomes. On n'utilise pas solve commepour les meme raison que pour
         % les racines de Legendre
         for i=2:nb_points
             zeros_tchebyschev(:, i-1) = [roots(sym2poly(poly_tchebyschev(i))); zeros(nb_points-i, 1)];
